requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
commonmark==0.9.1
cryptography==43.0.3
Django==5.2.4
django-cors-headers==4.7.0
django-dump-die==0.1.7
django-environ==0.12.0
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ExifRead==3.0.0
filelock==3.19.1
fsspec==2025.7.0
ggshield==1.42.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
MarkupSafe==3.0.2
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
oauthlib==3.2.2
packaging==25.0
phonenumbers==9.0.10
pillow==11.3.0
pillow-heif==0.16.0
platformdirs==3.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.22
pygitguardian==1.24.0
Pygments==2.19.2
PyJWT==2.6.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
PyYAML==6.0.2
qrcode==7.4.2
rawpy==0.25.1
redis==6.4.0
referencing==0.36.2
regex==2025.8.29
requests==2.32.4
rich==12.5.1
rpds-py==0.26.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
truststore==0.10.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
